math-engine/requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Mathematical computation
sympy==1.12
numpy==1.24.3
scipy==1.11.4
matplotlib==3.7.2

# Machine learning (for AI/ML math features)
torch==2.1.0
scikit-learn==1.3.0

# Data handling
pandas==2.0.3
pydantic==2.5.0

# Database and caching
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.23

# HTTP client for external APIs
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# AI/ML services
openai==1.3.7

# Environment and configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Image processing (for handwriting recognition)
Pillow==10.1.0

# Logging and monitoring
structlog==23.2.0
psutil==5.9.6

# Form data handling
python-multipart==0.0.6